refactoring(15): extract EarSection + useEAR from ResultsPage (high-risk seam)
Move the EAR (Experiment Artifact Repository) section out of the
ResultsPage container: renderEAR's JSX becomes EarSection.tsx (verbatim
body), and its 14 self-contained action-state vars (curate/publish +
publish.yaml-editor) move into a useEAR() hook. ear/earLoading/selectedId/
setEar/t are props; EarSection renders unconditionally where renderEAR()
was called, so it stays mounted across ear transitions and action-state
persistence is unchanged. ResultsPage 1327->922 lines.

Behavior-equivalence adversarially verified (3 skeptics: JSX byte-identity,
useEAR state fidelity, container wiring incl. mount-stability) — all
equivalent, no defects. typecheck 0 non-test errors; build ok; vitest
4 passed / 2 pre-existing.
