+++ ari-core/ari/viz/frontend/src/README.md
@@ -57,6 +57,7 @@ React/TypeScript source for the ARI `ari.viz` web dashboard — app entry, pages
       - `ResultsView.tsx` — leaf grades + rubric tree + negative-control display.
   - `Results/` — final run results and rubric scoring.
     - `README.md` — Results index.
+    - `EarSection.tsx` — Experiment Artifact Repository section (curate/publish/publish.yaml editor); extracted from ResultsPage renderEAR in req 15.
     - `index.ts` — barrel re-export.
     - `PublishYamlEditor.tsx` — per-checkpoint publish.yaml (EAR allowlist) editor; extracted from ResultsPage in req 03.
     - `resultHelpers.ts` — pure helpers + string formatters (tryParseJson, buildGradeMap, aggregateScore, format*Stage, etc.); extracted from resultSections in req 15.
@@ -64,6 +65,7 @@ React/TypeScript source for the ARI `ari.viz` web dashboard — app entry, pages
     - `ResultsPage.tsx` — results page container (state, data loading, layout).
     - `resultTypes.ts` — Results-page shared types (OrsRenderInput, RubricNode, LeafGrade, StageState); extracted from resultSections in req 15.
     - `RubricTreeVisualization.tsx` — D3 rubric tree with aggregated leaf scores.
+    - `useEAR.ts` — hook owning EarSection's curate/publish/publish.yaml-editor action state; extracted from ResultsPage in req 15.
   - `Settings/` — dashboard/run configuration page.
     - `README.md` — Settings index.
     - `index.ts` — barrel re-export.
